requirements.txt
tensorflow
tensorflow-datasets
matplotlib
numpy
pillow